requirements.txt
joblib==1.5.3
numpy==2.4.1
opencv-python==4.13.0.90
packaging==26.0
pillow==12.1.0
scikit-learn==1.8.0
scipy==1.17.0
setuptools==80.10.1
threadpoolctl==3.6.0
wheel==0.46.2
